pyproject.toml
[tool.poetry]
name = "hackforsrilanka_lightsoff"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.13"
psycopg2 = "^2.9.3"
django-environ = "^0.8.1"
gunicorn = "^20.1.0"
whitenoise = "^6.0.0"
twilio = "^7.8.1"
django-anymail = {extras = ["mailgun"], version = "^8.5"}
redis = "^4.2.2"
django-redis = "^5.2.0"
celery = "^5.2.6"
django-crispy-forms = "^1.14.0"
crispy-tailwind = "^0.5.0"
python-dateutil = "^2.8.2"
django-celery-beat = "^2.2.1"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
